requirements.txt
# REQUIREMENTS
python-dotenv == 0.12.0
Flask == 1.1.1
requests == 2.23.0
redis == 3.4.1